requirements.txt
discord-py-interactions~=4.0.0
PyDrive~=1.3.1
